+++ tests/converters/parsers_tests.py
@@ -40,6 +40,25 @@ class SimpleEventToVolumeTest(unittest.TestCase):
         )
 
 
+class SimpleEventToInstrumentTest(unittest.TestCase):
+    def setUp(self):
+        self.converter = music_converters.SimpleEventToInstrument()
+
+    def test_convert_with_attribute(self):
+        self.assertEqual(
+            self.converter(
+                music_events.NoteLike(instrument=music_parameters.constants.OBOE)
+            ),
+            music_parameters.constants.OBOE,
+        )
+
+    def test_convert_without_attribute(self):
+        self.assertEqual(
+            self.converter(core_events.SimpleEvent(10)),
+            None,
+        )
+
+
 class SimpleEventToPlayingIndicatorCollectionTest(unittest.TestCase):
     def setUp(self):
         self.converter = music_converters.SimpleEventToPlayingIndicatorCollection()